Moved substitution tests into their own dedicated test suite
